+++ package.json
@@ -4,13 +4,18 @@
   "description": "Shows Wikipedia article preview in a popup",
   "main": "dist/wikipedia-preview.production.js",
   "scripts": {
-    "test": "mocha --require @babel/register",
-    "test:watch": "mocha --require @babel/register -w",
+    "test": "mocha test/unit --require @babel/register",
+    "test:watch": "mocha test/unit --require @babel/register -w",
     "coverage": "nyc npm test",
     "build:dev": "webpack --mode development",
     "build:prod": "webpack --mode production",
     "build": "npm run build:dev && npm run build:prod",
-    "dev": "webpack-dev-server --mode development"
+    "dev": "webpack-dev-server --mode development",
+    "selenium-safari": "wdio test/e2e/config/wdio.safari.conf.js",
+    "selenium-edge": "wdio test/e2e/config/wdio.edge.conf.js",
+    "selenium-chrome": "wdio test/e2e/config/wdio.chrome.conf.js",
+    "selenium-ie": "wdio test/e2e/config/wdio.ie.conf.js",
+    "selenium-firefox": "wdio test/e2e/config/wdio.ff.conf.js"
   },
   "author": "Wikimedia Foundation",
   "license": "MIT",
@@ -18,12 +23,20 @@
     "@babel/core": "^7.8.3",
     "@babel/preset-env": "^7.8.3",
     "@babel/register": "^7.8.3",
+    "@wdio/cli": "5.23.0",
+    "@wdio/crossbrowsertesting-service": "^5.0.0",
+    "@wdio/local-runner": "5.23.0",
+    "@wdio/mocha-framework": "5.23.0",
+    "@wdio/spec-reporter": "5.23.0",
+    "axios": "^0.19.2",
     "babel-loader": "^8.0.6",
     "bufferutil": "^4.0.1",
     "canvas": "^2.6.1",
     "css-loader": "^3.4.2",
     "eslint": "^6.8.0",
     "eslint-loader": "^3.0.3",
+    "expect": "^26.0.1",
+    "expect-webdriverio": "^1.1.5",
     "jsdom": "^16.0.1",
     "less": "^3.10.3",
     "less-loader": "^5.0.0",
